@@ -1,8 +1,10 @@
 import 'dart:html' as html;
-import 'dart:ui';
+import 'dart:typed_data';
 
+import 'package:cross_file/cross_file.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
-import 'package:meta/meta.dart';
+import 'package:mime/mime.dart' show lookupMimeType;
 import 'package:share_plus_platform_interface/share_plus_platform_interface.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -13,8 +15,7 @@ class SharePlusPlugin extends SharePlatform {
     SharePlatform.instance = SharePlusPlugin();
   }
 
-  // ignore: prefer_typing_uninitialized_variables
-  final _navigator;
+  final html.Navigator _navigator;
 
   /// A constructor that allows tests to override the window object used by the plugin.
   SharePlusPlugin({@visibleForTesting html.Navigator? debugNavigator})
@@ -45,8 +46,8 @@ class SharePlusPlugin extends SharePlatform {
             .join('&'),
       );
 
-      if (await canLaunch(uri.toString())) {
-        await launch(uri.toString());
+      if (await canLaunchUrl(uri)) {
+        await launchUrl(uri);
       } else {
         throw Exception('Unable to share on web');
       }
@@ -62,6 +63,66 @@ class SharePlusPlugin extends SharePlatform {
     String? text,
     Rect? sharePositionOrigin,
   }) {
-    throw UnimplementedError('shareFiles() has not been implemented on Web.');
+    final files = <XFile>[];
+    for (var i = 0; i < paths.length; i++) {
+      files.add(XFile(paths[i], mimeType: mimeTypes?[i]));
+    }
+    return shareXFiles(
+      files,
+      subject: subject,
+      text: text,
+      sharePositionOrigin: sharePositionOrigin,
+    );
+  }
+
+  /// Share [XFile] objects.
+  ///
+  /// Remarks for the web implementation:
+  /// This uses the [Web Share API](https://web.dev/web-share/) if it's
+  /// available. Otherwise, uncaught Errors will be thrown.
+  /// See [Can I Use - Web Share API](https://caniuse.com/web-share) to
+  /// understand which browsers are supported. This builds on the
+  /// [`cross_file`](https://pub.dev/packages/cross_file) package.
+  @override
+  Future<ShareResult> shareXFiles(
+    List<XFile> files, {
+    String? subject,
+    String? text,
+    Rect? sharePositionOrigin,
+  }) async {
+    // See https://developer.mozilla.org/en-US/docs/Web/API/Navigator/share
+
+    final webFiles = <html.File>[];
+    for (final xFile in files) {
+      webFiles.add(await _fromXFile(xFile));
+    }
+    await _navigator.share({
+      if (subject?.isNotEmpty ?? false) 'title': subject,
+      if (text?.isNotEmpty ?? false) 'text': text,
+      if (webFiles.isNotEmpty) 'files': webFiles,
+    });
+
+    return _resultUnavailable;
+  }
+
+  static Future<html.File> _fromXFile(XFile file) async {
+    final bytes = await file.readAsBytes();
+    return html.File(
+      [ByteData.sublistView(bytes)],
+      file.name,
+      {
+        'type': file.mimeType ?? _mimeTypeForPath(file, bytes),
+      },
+    );
+  }
+
+  static String _mimeTypeForPath(XFile file, Uint8List bytes) {
+    return lookupMimeType(file.name, headerBytes: bytes) ??
+        'application/octet-stream';
   }
 }
+
+const _resultUnavailable = ShareResult(
+  'dev.fluttercommunity.plus/share/unavailable',
+  ShareResultStatus.unavailable,
+);
