@@ -2,15 +2,21 @@
 library share_plus_windows;
 
 import 'dart:ui';
-
+import 'package:cross_file/cross_file.dart';
 import 'package:url_launcher/url_launcher.dart';
+
+import 'package:share_plus_windows/src/version_helper.dart';
 import 'package:share_plus_platform_interface/share_plus_platform_interface.dart';
 
-/// The Windows implementation of SharePlatform.
+/// The fallback Windows implementation of [SharePlatform], for older Windows versions.
+///
 class ShareWindows extends SharePlatform {
-  /// Register this dart class as the platform implementation for linux
+  /// If the modern Share UI i.e. `DataTransferManager` is not available, then use this Dart class instead of platform specific implementation.
+  ///
   static void registerWith() {
-    SharePlatform.instance = ShareWindows();
+    if (!VersionHelper.instance.isWindows10RS5OrGreater) {
+      SharePlatform.instance = ShareWindows();
+    }
   }
 
   /// Share text.
@@ -34,8 +40,8 @@ class ShareWindows extends SharePlatform {
           .join('&'),
     );
 
-    if (await canLaunch(uri.toString())) {
-      await launch(uri.toString());
+    if (await canLaunchUrl(uri)) {
+      await launchUrl(uri);
     } else {
       throw Exception('Unable to share on windows');
     }
@@ -51,6 +57,20 @@ class ShareWindows extends SharePlatform {
     Rect? sharePositionOrigin,
   }) {
     throw UnimplementedError(
-        'shareFiles() has not been implemented on Windows.');
+      'shareFiles() is only available for Windows versions higher than 10.0.${VersionHelper.kWindows10RS5BuildNumber}.',
+    );
+  }
+
+  /// Share [XFile] objects with Result.
+  @override
+  Future<ShareResult> shareXFiles(
+    List<XFile> files, {
+    String? subject,
+    String? text,
+    Rect? sharePositionOrigin,
+  }) {
+    throw UnimplementedError(
+      'shareXFiles() is only available for Windows versions higher than 10.0.${VersionHelper.kWindows10RS5BuildNumber}.',
+    );
   }
 }
