@@ -7,8 +7,8 @@ import 'dart:async';
 import 'package:async/async.dart';
 import 'package:battery_plus/battery_plus.dart';
 import 'package:battery_plus_platform_interface/battery_plus_platform_interface.dart';
+import 'package:flutter_test/flutter_test.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-import 'package:test/test.dart';
 
 late StreamController<BatteryState> controller;
 
