@@ -28,11 +28,11 @@ public class PathMonitorConnectivityProvider: NSObject, ConnectivityProvider {
 
   override init() {
     super.init()
-    ensurePathMonitor()
+    _ = ensurePathMonitor()
   }
 
   public func start() {
-    ensurePathMonitor()
+    _ = ensurePathMonitor()
   }
 
   public func stop() {
@@ -40,6 +40,7 @@ public class PathMonitorConnectivityProvider: NSObject, ConnectivityProvider {
     _pathMonitor = nil
   }
 
+  @discardableResult
   private func ensurePathMonitor() -> NWPathMonitor {
     if (_pathMonitor == nil) {
       let pathMonitor = NWPathMonitor()
