@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:math' as math show sqrt;
 import 'base_device_info.dart';
 
 /// Information derived from `android.os.Build`.
@@ -29,8 +30,8 @@ class AndroidDeviceInfo implements BaseDeviceInfo {
     this.tags,
     this.type,
     this.isPhysicalDevice,
-    this.androidId,
     required List<String?> systemFeatures,
+    required this.displayMetrics,
   })  : supported32BitAbis = List<String?>.unmodifiable(supported32BitAbis),
         supported64BitAbis = List<String?>.unmodifiable(supported64BitAbis),
         supportedAbis = List<String?>.unmodifiable(supportedAbis),
@@ -93,9 +94,6 @@ class AndroidDeviceInfo implements BaseDeviceInfo {
   /// `false` if the application is running in an emulator, `true` otherwise.
   final bool? isPhysicalDevice;
 
-  /// The Android hardware device ID that is unique between the device + user and app signing.
-  final String? androidId;
-
   /// Describes what features are available on the current device.
   ///
   /// This can be used to check if the device has, for example, a front-facing
@@ -112,6 +110,9 @@ class AndroidDeviceInfo implements BaseDeviceInfo {
   /// https://developer.android.com/reference/android/content/pm/PackageManager
   final List<String?> systemFeatures;
 
+  /// Information about the current android display.
+  final AndroidDisplayMetrics displayMetrics;
+
   /// Serializes [AndroidDeviceInfo] to map.
   @override
   Map<String, dynamic> toMap() {
@@ -127,7 +128,6 @@ class AndroidDeviceInfo implements BaseDeviceInfo {
       'product': product,
       'display': display,
       'hardware': hardware,
-      'androidId': androidId,
       'bootloader': bootloader,
       'version': version.toMap(),
       'fingerprint': fingerprint,
@@ -137,6 +137,7 @@ class AndroidDeviceInfo implements BaseDeviceInfo {
       'isPhysicalDevice': isPhysicalDevice,
       'supported32BitAbis': supported32BitAbis,
       'supported64BitAbis': supported64BitAbis,
+      'displayMetrics': displayMetrics.toMap(),
     };
   }
 
@@ -163,8 +164,9 @@ class AndroidDeviceInfo implements BaseDeviceInfo {
       tags: map['tags'],
       type: map['type'],
       isPhysicalDevice: map['isPhysicalDevice'],
-      androidId: map['androidId'],
       systemFeatures: _fromList(map['systemFeatures'] ?? []),
+      displayMetrics: AndroidDisplayMetrics._fromMap(
+          map['displayMetrics']?.cast<String, dynamic>() ?? {}),
     );
   }
 
@@ -236,6 +238,65 @@ class AndroidBuildVersion {
       release: map['release'],
       sdkInt: map['sdkInt'],
       securityPatch: map['securityPatch'],
+    );
+  }
+}
+
+/// Details for the current display
+///
+/// See: https://developer.android.com/reference/android/util/DisplayMetrics
+class AndroidDisplayMetrics {
+  const AndroidDisplayMetrics._({
+    required this.widthPx,
+    required this.heightPx,
+    required this.xDpi,
+    required this.yDpi,
+  });
+
+  /// Gets the absolute width in pixels of the largest region of the display accessible to an app
+  /// in the current system state, without subtracting any window decor or applying scaling factors.
+  final double widthPx;
+
+  /// Gets the absolute height in pixels of the largest region of the display accessible to an app
+  /// in the current system state, without subtracting any window decor or applying scaling factors.
+  final double heightPx;
+
+  /// The exact physical pixels per inch of the screen in the X dimension.
+  final double xDpi;
+
+  /// The exact physical pixels per inch of the screen in the Y dimension.
+  final double yDpi;
+
+  /// The exact physical display width in inches.
+  double get widthInches => widthPx / xDpi;
+
+  /// The exact physical display height in inches.
+  double get heightInches => heightPx / yDpi;
+
+  /// The exact physical size in inches measured diagonally across the display.
+  double get sizeInches {
+    final width = widthInches;
+    final height = heightInches;
+    return math.sqrt((width * width) + (height * height));
+  }
+
+  /// Serializes [AndroidDisplayMetrics] to map.
+  Map<String, dynamic> toMap() {
+    return {
+      'widthPx': widthPx,
+      'heightPx': heightPx,
+      'xDpi': xDpi,
+      'yDpi': yDpi,
+    };
+  }
+
+  /// Deserializes from the map message received from the [MethodChannel].
+  static AndroidDisplayMetrics _fromMap(Map<String, dynamic> map) {
+    return AndroidDisplayMetrics._(
+      widthPx: map['widthPx'],
+      heightPx: map['heightPx'],
+      xDpi: map['xDpi'],
+      yDpi: map['yDpi'],
     );
   }
 }
