@@ -6,29 +6,23 @@
 
 #include <memory>
 
-#include "run_loop.h"
 #include "win32_window.h"
 
 // A window that does nothing but host a Flutter view.
 class FlutterWindow : public Win32Window {
-public:
-  // Creates a new FlutterWindow driven by the |run_loop|, hosting a
-  // Flutter view running |project|.
-  explicit FlutterWindow(RunLoop *run_loop,
-                         const flutter::DartProject &project);
+ public:
+  // Creates a new FlutterWindow hosting a Flutter view running |project|.
+  explicit FlutterWindow(const flutter::DartProject& project);
   virtual ~FlutterWindow();
 
-protected:
+ protected:
   // Win32Window:
   bool OnCreate() override;
   void OnDestroy() override;
   LRESULT MessageHandler(HWND window, UINT const message, WPARAM const wparam,
                          LPARAM const lparam) noexcept override;
 
-private:
-  // The run loop driving events for this window.
-  RunLoop *run_loop_;
-
+ private:
   // The project to run.
   flutter::DartProject project_;
 
@@ -36,4 +30,4 @@ private:
   std::unique_ptr<flutter::FlutterViewController> flutter_controller_;
 };
 
-#endif // RUNNER_FLUTTER_WINDOW_H_
+#endif  // RUNNER_FLUTTER_WINDOW_H_
