@@ -27,7 +27,7 @@ void EnableFullDpiSupportIfAvailable(HWND hwnd) {
     return;
   }
   auto enable_non_client_dpi_scaling =
-      reinterpret_cast<EnableNonClientDpiScaling *>(
+      reinterpret_cast<EnableNonClientDpiScaling*>(
           GetProcAddress(user32_module, "EnableNonClientDpiScaling"));
   if (enable_non_client_dpi_scaling != nullptr) {
     enable_non_client_dpi_scaling(hwnd);
@@ -35,15 +35,15 @@ void EnableFullDpiSupportIfAvailable(HWND hwnd) {
   }
 }
 
-} // namespace
+}  // namespace
 
 // Manages the Win32Window's window class registration.
 class WindowClassRegistrar {
-public:
+ public:
   ~WindowClassRegistrar() = default;
 
   // Returns the singleton registar instance.
-  static WindowClassRegistrar *GetInstance() {
+  static WindowClassRegistrar* GetInstance() {
     if (!instance_) {
       instance_ = new WindowClassRegistrar();
     }
@@ -52,23 +52,23 @@ public:
 
   // Returns the name of the window class, registering the class if it hasn't
   // previously been registered.
-  const wchar_t *GetWindowClass();
+  const wchar_t* GetWindowClass();
 
   // Unregisters the window class. Should only be called if there are no
   // instances of the window.
   void UnregisterWindowClass();
 
-private:
+ private:
   WindowClassRegistrar() = default;
 
-  static WindowClassRegistrar *instance_;
+  static WindowClassRegistrar* instance_;
 
   bool class_registered_ = false;
 };
 
-WindowClassRegistrar *WindowClassRegistrar::instance_ = nullptr;
+WindowClassRegistrar* WindowClassRegistrar::instance_ = nullptr;
 
-const wchar_t *WindowClassRegistrar::GetWindowClass() {
+const wchar_t* WindowClassRegistrar::GetWindowClass() {
   if (!class_registered_) {
     WNDCLASS window_class{};
     window_class.hCursor = LoadCursor(nullptr, IDC_ARROW);
@@ -93,18 +93,21 @@ void WindowClassRegistrar::UnregisterWindowClass() {
   class_registered_ = false;
 }
 
-Win32Window::Win32Window() { ++g_active_window_count; }
+Win32Window::Win32Window() {
+  ++g_active_window_count;
+}
 
 Win32Window::~Win32Window() {
   --g_active_window_count;
   Destroy();
 }
 
-bool Win32Window::CreateAndShow(const std::wstring &title, const Point &origin,
-                                const Size &size) {
+bool Win32Window::CreateAndShow(const std::wstring& title,
+                                const Point& origin,
+                                const Size& size) {
   Destroy();
 
-  const wchar_t *window_class =
+  const wchar_t* window_class =
       WindowClassRegistrar::GetInstance()->GetWindowClass();
 
   const POINT target_point = {static_cast<LONG>(origin.x),
@@ -127,18 +130,19 @@ bool Win32Window::CreateAndShow(const std::wstring &title, const Point &origin,
 }
 
 // static
-LRESULT CALLBACK Win32Window::WndProc(HWND const window, UINT const message,
+LRESULT CALLBACK Win32Window::WndProc(HWND const window,
+                                      UINT const message,
                                       WPARAM const wparam,
                                       LPARAM const lparam) noexcept {
   if (message == WM_NCCREATE) {
-    auto window_struct = reinterpret_cast<CREATESTRUCT *>(lparam);
+    auto window_struct = reinterpret_cast<CREATESTRUCT*>(lparam);
     SetWindowLongPtr(window, GWLP_USERDATA,
                      reinterpret_cast<LONG_PTR>(window_struct->lpCreateParams));
 
-    auto that = static_cast<Win32Window *>(window_struct->lpCreateParams);
+    auto that = static_cast<Win32Window*>(window_struct->lpCreateParams);
     EnableFullDpiSupportIfAvailable(window);
     that->window_handle_ = window;
-  } else if (Win32Window *that = GetThisFromHandle(window)) {
+  } else if (Win32Window* that = GetThisFromHandle(window)) {
     return that->MessageHandler(window, message, wparam, lparam);
   }
 
@@ -146,41 +150,44 @@ LRESULT CALLBACK Win32Window::WndProc(HWND const window, UINT const message,
 }
 
 LRESULT
-Win32Window::MessageHandler(HWND hwnd, UINT const message, WPARAM const wparam,
+Win32Window::MessageHandler(HWND hwnd,
+                            UINT const message,
+                            WPARAM const wparam,
                             LPARAM const lparam) noexcept {
   switch (message) {
-  case WM_DESTROY:
-    window_handle_ = nullptr;
-    Destroy();
-    if (quit_on_close_) {
-      PostQuitMessage(0);
+    case WM_DESTROY:
+      window_handle_ = nullptr;
+      Destroy();
+      if (quit_on_close_) {
+        PostQuitMessage(0);
+      }
+      return 0;
+
+    case WM_DPICHANGED: {
+      auto newRectSize = reinterpret_cast<RECT*>(lparam);
+      LONG newWidth = newRectSize->right - newRectSize->left;
+      LONG newHeight = newRectSize->bottom - newRectSize->top;
+
+      SetWindowPos(hwnd, nullptr, newRectSize->left, newRectSize->top, newWidth,
+                   newHeight, SWP_NOZORDER | SWP_NOACTIVATE);
+
+      return 0;
     }
-    return 0;
-
-  case WM_DPICHANGED: {
-    auto newRectSize = reinterpret_cast<RECT *>(lparam);
-    LONG newWidth = newRectSize->right - newRectSize->left;
-    LONG newHeight = newRectSize->bottom - newRectSize->top;
-
-    SetWindowPos(hwnd, nullptr, newRectSize->left, newRectSize->top, newWidth,
-                 newHeight, SWP_NOZORDER | SWP_NOACTIVATE);
-
-    return 0;
-  }
-  case WM_SIZE:
-    RECT rect = GetClientArea();
-    if (child_content_ != nullptr) {
-      // Size and position the child window.
-      MoveWindow(child_content_, rect.left, rect.top, rect.right - rect.left,
-                 rect.bottom - rect.top, TRUE);
+    case WM_SIZE: {
+      RECT rect = GetClientArea();
+      if (child_content_ != nullptr) {
+        // Size and position the child window.
+        MoveWindow(child_content_, rect.left, rect.top, rect.right - rect.left,
+                   rect.bottom - rect.top, TRUE);
+      }
+      return 0;
     }
-    return 0;
 
-  case WM_ACTIVATE:
-    if (child_content_ != nullptr) {
-      SetFocus(child_content_);
-    }
-    return 0;
+    case WM_ACTIVATE:
+      if (child_content_ != nullptr) {
+        SetFocus(child_content_);
+      }
+      return 0;
   }
 
   return DefWindowProc(window_handle_, message, wparam, lparam);
@@ -198,8 +205,8 @@ void Win32Window::Destroy() {
   }
 }
 
-Win32Window *Win32Window::GetThisFromHandle(HWND const window) noexcept {
-  return reinterpret_cast<Win32Window *>(
+Win32Window* Win32Window::GetThisFromHandle(HWND const window) noexcept {
+  return reinterpret_cast<Win32Window*>(
       GetWindowLongPtr(window, GWLP_USERDATA));
 }
 
@@ -220,7 +227,9 @@ RECT Win32Window::GetClientArea() {
   return frame;
 }
 
-HWND Win32Window::GetHandle() { return window_handle_; }
+HWND Win32Window::GetHandle() {
+  return window_handle_;
+}
 
 void Win32Window::SetQuitOnClose(bool quit_on_close) {
   quit_on_close_ = quit_on_close;
