@@ -11,7 +11,7 @@
 // inherited from by classes that wish to specialize with custom
 // rendering and input handling
 class Win32Window {
-public:
+ public:
   struct Point {
     unsigned int x;
     unsigned int y;
@@ -34,8 +34,9 @@ public:
   // consistent size to will treat the width height passed in to this function
   // as logical pixels and scale to appropriate for the default monitor. Returns
   // true if the window was created successfully.
-  bool CreateAndShow(const std::wstring &title, const Point &origin,
-                     const Size &size);
+  bool CreateAndShow(const std::wstring& title,
+                     const Point& origin,
+                     const Size& size);
 
   // Release OS resources associated with window.
   void Destroy();
@@ -53,11 +54,12 @@ public:
   // Return a RECT representing the bounds of the current client area.
   RECT GetClientArea();
 
-protected:
+ protected:
   // Processes and route salient window messages for mouse handling,
   // size change and DPI. Delegates handling of these to member overloads that
   // inheriting classes can handle.
-  virtual LRESULT MessageHandler(HWND window, UINT const message,
+  virtual LRESULT MessageHandler(HWND window,
+                                 UINT const message,
                                  WPARAM const wparam,
                                  LPARAM const lparam) noexcept;
 
@@ -68,7 +70,7 @@ protected:
   // Called when Destroy is called.
   virtual void OnDestroy();
 
-private:
+ private:
   friend class WindowClassRegistrar;
 
   // OS callback called by message pump. Handles the WM_NCCREATE message which
@@ -76,12 +78,13 @@ private:
   // non-client DPI scaling so that the non-client area automatically
   // responsponds to changes in DPI. All other messages are handled by
   // MessageHandler.
-  static LRESULT CALLBACK WndProc(HWND const window, UINT const message,
+  static LRESULT CALLBACK WndProc(HWND const window,
+                                  UINT const message,
                                   WPARAM const wparam,
                                   LPARAM const lparam) noexcept;
 
   // Retrieves a class instance pointer for |window|
-  static Win32Window *GetThisFromHandle(HWND const window) noexcept;
+  static Win32Window* GetThisFromHandle(HWND const window) noexcept;
 
   bool quit_on_close_ = false;
 
@@ -92,4 +95,4 @@ private:
   HWND child_content_ = nullptr;
 };
 
-#endif // RUNNER_WIN32_WINDOW_H_
+#endif  // RUNNER_WIN32_WINDOW_H_
