@@ -4,11 +4,27 @@
 
 #import "FLTSharePlusPlugin.h"
 #import "LinkPresentation/LPLinkMetadata.h"
+#import "LinkPresentation/LPMetadataProvider.h"
 
 static NSString *const PLATFORM_CHANNEL = @"dev.fluttercommunity.plus/share";
 
 static UIViewController *RootViewController() {
-  return [UIApplication sharedApplication].keyWindow.rootViewController;
+  if (@available(iOS 13, *)) { // UIApplication.keyWindow is deprecated
+    NSSet *scenes = [[UIApplication sharedApplication] connectedScenes];
+    for (UIScene *scene in scenes) {
+      if ([scene isKindOfClass:[UIWindowScene class]]) {
+        NSArray *windows = ((UIWindowScene *)scene).windows;
+        for (UIWindow *window in windows) {
+          if (window.isKeyWindow) {
+            return window.rootViewController;
+          }
+        }
+      }
+    }
+    return nil;
+  } else {
+    return [UIApplication sharedApplication].keyWindow.rootViewController;
+  }
 }
 
 static UIViewController *
@@ -75,8 +91,13 @@ TopViewControllerForViewController(UIViewController *viewController) {
 
 - (instancetype)initWithSubject:(NSString *)subject
                            text:(NSString *)text NS_DESIGNATED_INITIALIZER;
+
 - (instancetype)initWithFile:(NSString *)path
                     mimeType:(NSString *)mimeType NS_DESIGNATED_INITIALIZER;
+
+- (instancetype)initWithFile:(NSString *)path
+                    mimeType:(NSString *)mimeType
+                     subject:(NSString *)subject NS_DESIGNATED_INITIALIZER;
 
 - (instancetype)init
     __attribute__((unavailable("Use initWithSubject:text: instead")));
@@ -108,9 +129,23 @@ TopViewControllerForViewController(UIViewController *viewController) {
   return self;
 }
 
+- (instancetype)initWithFile:(NSString *)path
+                    mimeType:(NSString *)mimeType
+                     subject:(NSString *)subject {
+  self = [super init];
+  if (self) {
+    _path = path;
+    _mimeType = mimeType;
+    _subject = [subject isKindOfClass:NSNull.class] ? @"" : subject;
+  }
+  return self;
+}
+
 - (id)activityViewControllerPlaceholderItem:
     (UIActivityViewController *)activityViewController {
-  return @"";
+  return [self
+      activityViewController:activityViewController
+         itemForActivityType:@"dev.fluttercommunity.share_plus.placeholder"];
 }
 
 - (id)activityViewController:(UIActivityViewController *)activityViewController
@@ -119,13 +154,18 @@ TopViewControllerForViewController(UIViewController *viewController) {
     return _text;
   }
 
-  if ([_mimeType hasPrefix:@"image/"]) {
+  // If the shared file is an image return an UIImage for the placeholder
+  // to show a preview.
+  if ([activityType
+          isEqualToString:@"dev.fluttercommunity.share_plus.placeholder"] &&
+      [_mimeType hasPrefix:@"image/"]) {
     UIImage *image = [UIImage imageWithContentsOfFile:_path];
     return image;
-  } else {
-    NSURL *url = [NSURL fileURLWithPath:_path];
-    return url;
   }
+
+  // Return an NSURL for the real share to conserve the file name
+  NSURL *url = [NSURL fileURLWithPath:_path];
+  return url;
 }
 
 - (NSString *)activityViewController:
@@ -158,11 +198,42 @@ TopViewControllerForViewController(UIViewController *viewController) {
     (UIActivityViewController *)activityViewController
     API_AVAILABLE(macos(10.15), ios(13.0), watchos(6.0)) {
   LPLinkMetadata *metadata = [[LPLinkMetadata alloc] init];
-  if ([_text length] > 0) {
-    metadata.title = _text;
-  } else if ([_subject length] > 0) {
+
+  if ([_subject length] > 0) {
     metadata.title = _subject;
+  } else if ([_text length] > 0) {
+    metadata.title = _text;
   }
+
+  if (_path) {
+    NSString *extesnion = [_path pathExtension];
+
+    unsigned long long rawSize = (
+        [[[NSFileManager defaultManager] attributesOfItemAtPath:_path
+                                                          error:nil] fileSize]);
+    NSString *readableSize = [NSByteCountFormatter
+        stringFromByteCount:rawSize
+                 countStyle:NSByteCountFormatterCountStyleFile];
+
+    NSString *description = @"";
+
+    if (![extesnion isEqualToString:@""]) {
+      description =
+          [description stringByAppendingString:[extesnion uppercaseString]];
+      description = [description stringByAppendingString:@" • "];
+      description = [description stringByAppendingString:readableSize];
+    } else {
+      description = [description stringByAppendingString:readableSize];
+    }
+
+    // https://stackoverflow.com/questions/60563773/ios-13-share-sheet-changing-subtitle-item-description
+    metadata.originalURL = [NSURL fileURLWithPath:description];
+    if (_mimeType && [_mimeType hasPrefix:@"image/"]) {
+      metadata.imageProvider = [[NSItemProvider alloc]
+          initWithObject:[UIImage imageWithContentsOfFile:_path]];
+    }
+  }
+
   return metadata;
 }
 
@@ -203,13 +274,22 @@ TopViewControllerForViewController(UIViewController *viewController) {
             return;
           }
 
+          UIViewController *rootViewController = RootViewController();
+          if (!rootViewController) {
+            result([FlutterError errorWithCode:@"error"
+                                       message:@"No root view controller found"
+                                       details:nil]);
+            return;
+          }
           UIViewController *topViewController =
-              TopViewControllerForViewController(RootViewController());
+              TopViewControllerForViewController(rootViewController);
+
           [self shareText:shareText
                      subject:shareSubject
               withController:topViewController
                     atSource:originRect
-                    toResult:withResult ? result : nil];
+                    toResult:result
+                  withResult:withResult];
           if (!withResult)
             result(nil);
         } else if ([@"shareFiles" isEqualToString:call.method] ||
@@ -235,15 +315,23 @@ TopViewControllerForViewController(UIViewController *viewController) {
             }
           }
 
+          UIViewController *rootViewController = RootViewController();
+          if (!rootViewController) {
+            result([FlutterError errorWithCode:@"error"
+                                       message:@"No root view controller found"
+                                       details:nil]);
+            return;
+          }
           UIViewController *topViewController =
-              TopViewControllerForViewController(RootViewController());
+              TopViewControllerForViewController(rootViewController);
           [self shareFiles:paths
                 withMimeType:mimeTypes
                  withSubject:subject
                     withText:text
               withController:topViewController
                     atSource:originRect
-                    toResult:withResult ? result : nil];
+                    toResult:result
+                  withResult:withResult];
           if (!withResult)
             result(nil);
         } else {
@@ -256,7 +344,8 @@ TopViewControllerForViewController(UIViewController *viewController) {
        withSubject:(NSString *)subject
     withController:(UIViewController *)controller
           atSource:(CGRect)origin
-          toResult:(FlutterResult)result {
+          toResult:(FlutterResult)result
+        withResult:(BOOL)withResult {
   UIActivityViewSuccessController *activityViewController =
       [[UIActivityViewSuccessController alloc] initWithActivityItems:shareItems
                                                applicationActivities:nil];
@@ -268,10 +357,32 @@ TopViewControllerForViewController(UIViewController *viewController) {
 
   activityViewController.popoverPresentationController.sourceView =
       controller.view;
+  BOOL isCoordinateSpaceOfSourceView =
+      CGRectContainsRect(controller.view.frame, origin);
+
+  // If device is e.g. an iPad then hasPopoverPresentationController is true
+  BOOL hasPopoverPresentationController =
+      [activityViewController popoverPresentationController] != NULL;
+  if (hasPopoverPresentationController &&
+      (!isCoordinateSpaceOfSourceView || CGRectIsEmpty(origin))) {
+    NSString *sharePositionIssue = [NSString
+        stringWithFormat:
+            @"sharePositionOrigin: argument must be set, %@ must be non-zero "
+            @"and within coordinate space of source view: %@",
+            NSStringFromCGRect(origin),
+            NSStringFromCGRect(controller.view.bounds)];
+
+    result([FlutterError errorWithCode:@"error"
+                               message:sharePositionIssue
+                               details:nil]);
+    return;
+  }
+
   if (!CGRectIsEmpty(origin)) {
     activityViewController.popoverPresentationController.sourceRect = origin;
   }
-  if (result) {
+
+  if (withResult) {
     UIActivityViewSuccessCompanion *companion =
         [[UIActivityViewSuccessCompanion alloc] initWithResult:result];
     activityViewController.companion = companion;
@@ -291,16 +402,16 @@ TopViewControllerForViewController(UIViewController *viewController) {
            subject:(NSString *)subject
     withController:(UIViewController *)controller
           atSource:(CGRect)origin
-          toResult:(FlutterResult)result {
-  NSObject *data = [[NSURL alloc] initWithString:shareText];
-  if (data == nil) {
-    data = [[SharePlusData alloc] initWithSubject:subject text:shareText];
-  }
+          toResult:(FlutterResult)result
+        withResult:(BOOL)withResult {
+  NSObject *data = [[SharePlusData alloc] initWithSubject:subject
+                                                     text:shareText];
   [self share:@[ data ]
          withSubject:subject
       withController:controller
             atSource:origin
-            toResult:result];
+            toResult:result
+          withResult:withResult];
 }
 
 + (void)shareFiles:(NSArray *)paths
@@ -309,36 +420,28 @@ TopViewControllerForViewController(UIViewController *viewController) {
           withText:(NSString *)text
     withController:(UIViewController *)controller
           atSource:(CGRect)origin
-          toResult:(FlutterResult)result {
+          toResult:(FlutterResult)result
+        withResult:(BOOL)withResult {
   NSMutableArray *items = [[NSMutableArray alloc] init];
-
-  if (text || subject) {
-    [items addObject:[[SharePlusData alloc] initWithSubject:subject text:text]];
-  }
 
   for (int i = 0; i < [paths count]; i++) {
     NSString *path = paths[i];
-    NSString *pathExtension = [path pathExtension];
     NSString *mimeType = mimeTypes[i];
-    if ([pathExtension.lowercaseString isEqualToString:@"jpg"] ||
-        [pathExtension.lowercaseString isEqualToString:@"jpeg"] ||
-        [pathExtension.lowercaseString isEqualToString:@"png"] ||
-        [mimeType.lowercaseString isEqualToString:@"image/jpg"] ||
-        [mimeType.lowercaseString isEqualToString:@"image/jpeg"] ||
-        [mimeType.lowercaseString isEqualToString:@"image/png"]) {
-      UIImage *image = [UIImage imageWithContentsOfFile:path];
-      [items addObject:image];
-    } else {
-      NSURL *fileUrl = [NSURL fileURLWithPath:path];
-      [items addObject:fileUrl];
-    }
+    [items addObject:[[SharePlusData alloc] initWithFile:path
+                                                mimeType:mimeType
+                                                 subject:subject]];
+  }
+  if (text != nil) {
+    NSObject *data = [[SharePlusData alloc] initWithSubject:subject text:text];
+    [items addObject:data];
   }
 
   [self share:items
          withSubject:subject
       withController:controller
             atSource:origin
-            toResult:result];
+            toResult:result
+          withResult:withResult];
 }
 
 @end
