@@ -3,16 +3,15 @@
 // found in the LICENSE file.
 
 import 'dart:io';
-import 'dart:ui';
 
+import 'package:cross_file/cross_file.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart'
     show TestDefaultBinaryMessengerBinding, TestWidgetsFlutterBinding;
 import 'package:mockito/mockito.dart';
-import 'package:share_plus_platform_interface/share_plus_platform_interface.dart';
 import 'package:share_plus_platform_interface/method_channel/method_channel_share.dart';
+import 'package:share_plus_platform_interface/share_plus_platform_interface.dart';
 import 'package:test/test.dart';
-
-import 'package:flutter/services.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -29,6 +28,7 @@ void main() {
             (MethodCall call) async {
       // The explicit type can be void as the only method call has a return type of void.
       await mockChannel.invokeMethod<void>(call.method, call.arguments);
+      return null;
     });
   });
 
@@ -131,6 +131,26 @@ void main() {
           'originHeight': 4.0,
         },
       ));
+
+      await sharePlatform.shareXFiles(
+        [XFile(fd.path)],
+        subject: 'some subject to share',
+        text: 'some text to share',
+        sharePositionOrigin: const Rect.fromLTWH(1.0, 2.0, 3.0, 4.0),
+      );
+      verify(mockChannel.invokeMethod<void>(
+        'shareFilesWithResult',
+        <String, dynamic>{
+          'paths': [fd.path],
+          'mimeTypes': ['image/png'],
+          'subject': 'some subject to share',
+          'text': 'some text to share',
+          'originX': 1.0,
+          'originY': 2.0,
+          'originWidth': 3.0,
+          'originHeight': 4.0,
+        },
+      ));
     });
   });
 
@@ -141,6 +161,10 @@ void main() {
     );
     expect(
       () => SharePlatform.instance.shareFilesWithResult(['']),
+      throwsA(const TypeMatcher<AssertionError>()),
+    );
+    expect(
+      () => sharePlatform.shareXFiles([XFile('')]),
       throwsA(const TypeMatcher<AssertionError>()),
     );
     verifyZeroInteractions(mockChannel);
@@ -159,6 +183,12 @@ void main() {
         'paths': [fd.path],
         'mimeTypes': ['image/png'],
       }));
+
+      await sharePlatform.shareXFiles([XFile(fd.path)]);
+      verify(mockChannel.invokeMethod('shareFilesWithResult', <String, dynamic>{
+        'paths': [fd.path],
+        'mimeTypes': ['image/png'],
+      }));
     });
   });
 
@@ -172,6 +202,12 @@ void main() {
 
       await SharePlatform.instance
           .shareFilesWithResult([fd.path], mimeTypes: ['*/*']);
+      verify(mockChannel.invokeMethod('shareFilesWithResult', <String, dynamic>{
+        'paths': [fd.path],
+        'mimeTypes': ['*/*'],
+      }));
+
+      await sharePlatform.shareXFiles([XFile(fd.path, mimeType: '*/*')]);
       verify(mockChannel.invokeMethod('shareFilesWithResult', <String, dynamic>{
         'paths': [fd.path],
         'mimeTypes': ['*/*'],
@@ -216,6 +252,14 @@ void main() {
     await withFile('tempfile-83649e.png', (File fd) async {
       await sharePlatform.shareFilesWithResult([fd.path]);
       verify(mockChannel.invokeMethod('shareFiles', <String, dynamic>{
+        'paths': [fd.path],
+        'mimeTypes': ['image/png'],
+      }));
+    });
+
+    await withFile('tempfile-83649e.png', (File fd) async {
+      await sharePlatform.shareXFiles([XFile(fd.path)]);
+      verify(mockChannel.invokeMethod('shareFilesWithResult', <String, dynamic>{
         'paths': [fd.path],
         'mimeTypes': ['image/png'],
       }));
